Added a dictionary for the names
Previously, the main data structure used to store the initial spendings
was a dictionary. It used the name of the people as keys and their
initial spendings as values.
This proved to be inconvenient when implementing the actual algorithm.
Therefore, the payments are now just stored in a list, which allows for
easy indexing. A new dictionary, named name_dict has been added, which
guarantees, that the indices can still be linked to the names. So now
the spending of a person with the name `name` can be accessed via
start_accounts[name_dict[name]]