Implemented the actual algorithm
The main algorithm of Sinabs has been implemented. It is realized in
the generate_payments function and works in the following way:
First, the maximum, minimum and average of the accounts is found.
Next the algorithm checks, whether the difference between the maximum
and the average or the difference between the minimum and the average
is larger. The smaller of the two is set to be, what the minimum
spender has to pay to the maximum spender. Afterwards the new min,
max and average are evaluated again and the process is repeated, until
all payments left are less than one cent (0.01).
The results are then printed to stdout.